+++ lib/swiss_knife.dart
@@ -4,6 +4,7 @@ export 'src/collections.dart';
 export 'src/data.dart';
 export 'src/date.dart';
 export 'src/events.dart';
+export 'src/expando_with_finalizer.dart';
 export 'src/flat_hash_map.dart';
 export 'src/json.dart';
 export 'src/loader.dart';